Refactored monitor page for readability
